wpt_babel/content-security-policy/navigate-to/child-navigates-parent-blocked.sub.html

<!DOCTYPE html>
<head>
 <script src="/resources/testharness.js">
 </script>
 <script src="/resources/testharnessreport.js">
 </script>
</head>
<body>
 <script>
  "use strict";

var t = async_test("Test that the child can't navigate the parent because the relevant policy belongs to the navigation initiator (in this case the child which has the policy `navigate-to 'none'`)");
window.onmessage = t.step_func_done(function (e) {
  assert_equals(e.data.result, 'fail');
  assert_equals(e.data.violatedDirective, 'navigate-to');
});
 </script>
 <iframe srcdoc="&lt;iframe src='support/navigate_parent.sub.html?csp=navigate-to%20%27none%27&amp;report_id={{$id:uuid()}}'&gt;">
 </iframe>
 <script async="" defer="" src="../support/checkReport.sub.js?reportField=violated-directive&amp;reportValue=navigate-to%20%27none%27&amp;reportID={{$id}}">
 </script>
</body>
